requirements.txt
langchain
langchain_community
groq
python-dotenv
streamlit==1.34.0
pandas==1.5.3
pymongo==4.7.2
langchain-groq==0.1.4
python-dotenv==1.0.1
pandasai==2.0.41
streamlit-chat==0.1.1
